pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {}}

[project]
name = "email-processor"
version = "8.0.1"
description = "Email attachment processor with IMAP and SMTP support"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Valerii Kholodilin", email = "kholodilin.valerii@gmail.com"}
]
keywords = ["email", "imap", "smtp", "attachment", "processor", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "pyyaml>=6.0",
    "keyring>=24.0",
    "structlog>=24.0.0",
    "tqdm>=4.66.0",
    "cryptography>=40.0.0",  # Note: 32-bit Python may require older version (40.0.2), 64-bit Python supports >=41.0.0
]

[project.urls]
"Homepage" = "https://github.com/KHolodilin/python-email-automation-processor"
"Bug Reports" = "https://github.com/KHolodilin/python-email-automation-processor/issues"
"Source" = "https://github.com/KHolodilin/python-email-automation-processor"
"Documentation" = "https://github.com/KHolodilin/python-email-automation-processor#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
email-processor = "email_processor.__main__:main"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "email_processor.imap.mock_client",
]
ignore_errors = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "email_processor.security.encryption",
    "cryptography.*",
]
ignore_missing_imports = true
warn_return_any = false
disable_error_code = ["no-any-return"]

[[tool.mypy.overrides]]
module = "email_processor.security.fingerprint"
ignore_errors = true
disable_error_code = ["attr-defined"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=email_processor",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["email_processor"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/mock_client.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
